chore(DataMapper): Split MappingLink component
fix: https://github.com/KaotoIO/kaoto/issues/2396
