+++ packages/ui/src/components/View/MappingLink.tsx
@@ -1,12 +1,11 @@
-import { CSSProperties, FunctionComponent, useCallback, useEffect, useRef, useState } from 'react';
-import { useCanvas } from '../../hooks/useCanvas';
-import { Circle, LinePath } from '@visx/shape';
 import { curveMonotoneX } from '@visx/curve';
+import { Circle, LinePath } from '@visx/shape';
+import { CSSProperties, FunctionComponent, useCallback, useState } from 'react';
+import { useCanvas } from '../../hooks/useCanvas';
 import { useMappingLinks } from '../../hooks/useMappingLinks';
 import { LineProps } from '../../models/datamapper';
-import { MappingLinksService } from '../../services/mapping-links.service';
 
-const MappingLink: FunctionComponent<LineProps> = ({
+export const MappingLink: FunctionComponent<LineProps> = ({
   x1,
   y1,
   x2,
@@ -65,59 +64,5 @@ const MappingLink: FunctionComponent<LineProps> = ({
       />
       <Circle r={dotRadius} cx={x2} cy={y2} />
     </>
-  );
-};
-
-export const MappingLinksContainer: FunctionComponent = () => {
-  const [lineCoordList, setLineCoordList] = useState<LineProps[]>([]);
-  const { getNodeReference } = useCanvas();
-  const { getMappingLinks } = useMappingLinks();
-  const svgRef = useRef<SVGSVGElement>(null);
-
-  const refreshLinks = useCallback(() => {
-    const links = getMappingLinks();
-    const answer = MappingLinksService.calculateMappingLinkCoordinates(links, svgRef, getNodeReference);
-    setLineCoordList(answer);
-  }, [getMappingLinks, getNodeReference]);
-
-  useEffect(() => {
-    refreshLinks();
-    window.addEventListener('resize', refreshLinks);
-    window.addEventListener('scroll', refreshLinks);
-    return () => {
-      window.removeEventListener('resize', refreshLinks);
-      window.removeEventListener('scroll', refreshLinks);
-    };
-  }, [refreshLinks]);
-
-  return (
-    <svg
-      ref={svgRef}
-      data-testid="mapping-links"
-      style={{
-        position: 'absolute',
-        pointerEvents: 'none',
-        top: 0,
-        left: 0,
-        width: '100%',
-        height: '100%',
-      }}
-    >
-      <g z={0}>
-        {lineCoordList.map((lineProps, index) => (
-          <MappingLink
-            key={index}
-            svgRef={svgRef}
-            x1={lineProps.x1}
-            y1={lineProps.y1}
-            x2={lineProps.x2}
-            y2={lineProps.y2}
-            sourceNodePath={lineProps.sourceNodePath}
-            targetNodePath={lineProps.targetNodePath}
-            isSelected={lineProps.isSelected}
-          />
-        ))}
-      </g>
-    </svg>
   );
 };
